Support ruby 3
The delegation in #method_missing is breaking with ruby 3 because
for ruby 3 the delegation should explicitly delegate both positional
and keyword arguments. I'm using ruby2_keywords to keep the
old behavior until the support of ruby < 3 is dropped.

The method ruby2_keywords is present only in rubies that are >= 2.7
and that is why the check if respond_to?(:ruby2_keywords, true) is needed.